+++ src/shared/components/CallToActionButton/CallToActionButton.tsx
@@ -1,10 +1,20 @@
-import React, { FC } from 'react'
+import { To } from 'history'
+import React, { FC, MouseEvent, ReactNode } from 'react'
 
 import { SvgGlyphChevronRight } from '@/shared/icons'
 import { getLinkPropsFromTo } from '@/utils/button'
 
 import { BodyWrapper, ContentWrapper, IconWrapper, StyledContainer } from './CallToActionButton.style'
-import { CallToActionButtonProps } from './CallToActionButton.types'
+
+export type ColorVariants = 'red' | 'green' | 'yellow' | 'blue'
+
+export type CallToActionButtonProps = {
+  to?: To
+  onClick?: (event: MouseEvent<HTMLButtonElement>) => void
+  icon?: ReactNode
+  colorVariant?: ColorVariants
+  label: string
+}
 
 export const CallToActionButton: FC<CallToActionButtonProps> = ({
   to,
