+++ src/shared/components/CallToActionButton/index.ts
@@ -1,3 +1,3 @@
 export * from './CallToActionButton'
 export { CallToActionWrapper } from './CallToActionButton.style'
-export type { CallToActionButtonProps } from './CallToActionButton.types'
+export type { CallToActionButtonProps } from './CallToActionButton'
