+++ src/views/viewer/PopularView/PopularView.tsx
@@ -18,7 +18,7 @@ export const PopularView: FC = () => (
       <CallToActionButton label="Home" to={absoluteRoutes.viewer.index()} colorVariant="yellow" icon={<SvgNavHome />} />
       <CallToActionButton
         label="Browse channels"
-        to={absoluteRoutes.viewer.browseChannels()}
+        to={absoluteRoutes.viewer.channels()}
         colorVariant="blue"
         icon={<SvgNavChannels />}
       />
